+++ src/components/ProductCard.js
@@ -7,7 +7,7 @@ import {
   Button,
 } from "@mui/material";
 import { ShoppingCartOutlined } from "@mui/icons-material";
-import productImg from "../assets/productImg/pexels-ylanite-koppens-1152665.jpg";
+import productImg from "../assets/desk-covered-with-school-supplies-colorful-alarm-clocks.jpg";
 import { useDispatch } from "react-redux";
 import { addItemToBasket } from "../redux/basketSlice";
 
@@ -23,17 +23,18 @@ const ProductCard = ({ product }) => {
   return (
     <Card>
       <Box sx={{ pt: "100%", position: "relative" }}>
-        <Box
-          sx={{
-            width: "100%",
-            height: "100%",
-            position: "absolute",
-            top: 0,
-          }}
-          component="img"
+        <img
           src={imageUrl || productImg}
           alt={name}
           loading="lazy"
+          style={{
+            width: "100%",
+            height: "100%",
+            objectFit: "cover", 
+            position: "absolute",
+            top: 0,
+            left: 0,
+          }}
         />
       </Box>
